Fix: Move reservation directory
